requirements.txt
altair==4.2.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
scikit_learn==1.0.2
streamlit==1.9.0
yfinance==0.1.70
